1) Usunięcie widgetów EMG, GRF i Video 2) Przeniesienie wizualizatora wykresów do Chart 3) Dodanie wizualizatora Video 4) Wiele parserów może parsować ten sam plik 5) Parser nie musi wszystkim ObjectWrapperom nadawać niezerowe wartości 6) Liczne poprawki w UI 7) Dodanie ikon do wizualizatorów 8) Dodanie paczki ikon Icons.qrc oraz obsługa plików qrc przez CMake'a; dzięki temu można:   - w designerze widzieć ikony   - wkompilować ikonki w źródło, dzięki czemu szybciej się ładują
git-svn-id: https://localhost:8443/svn/hm-mde-1@401 fca3399c-2208-499e-aaa1-2cf7fb07c25e

+++ trunk/edr/include/core/PluginLoader.h
@@ -44,19 +44,19 @@ public:
     void load();
 
     //! \return
-    size_t getNumPlugins() const
+    int getNumPlugins() const
     {
-        return plugins.size();
+        return static_cast<int>(plugins.size());
     }
     //! 
     //! \param idx
-    PluginPtr getPlugin(size_t idx)
+    PluginPtr getPlugin(int idx)
     {
         return plugins[idx];
     }
     //! 
     //! \param idx
-    const PluginPtr getPlugin(size_t idx) const
+    const PluginPtr getPlugin(int idx) const
     {
         return plugins[idx];
     }
